extract InheritFromMappingAndSpecification to extension metnod
