+++ src/AppForeach.Framework.FluentValidation/AppForeach.Framework.FluentValidation.csproj
@@ -11,6 +11,7 @@
   </ItemGroup>	
 
   <ItemGroup>
+    <ProjectReference Include="..\AppForeach.Framework.Automapper\AppForeach.Framework.Automapper.csproj" />
     <ProjectReference Include="..\AppForeach.Framework\AppForeach.Framework.csproj" />
   </ItemGroup>
